更改文件结构
