+++ src/platform/base/error/error.js
