+++ src/platform/base/error/error.ts
