+++ src/platform/base/extend/extend.js
