+++ src/platform/base/extend/extend.ts
