+++ src/platform/base/log/log.js
