+++ src/platform/base/log/log.ts
