+++ src/platform/ipc/handlers/ipc.handler.js
@@ -3,7 +3,7 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.MainHandler = void 0;
 const electron_1 = require("electron");
 const ipc_events_1 = require("./ipc.events");
-const log_1 = require("../base/log");
+const log_1 = require("../base/log/log");
 var MainHandler;
 (function (MainHandler) {
     function initBind(mainWindow) {
